+++ src/Campaigns/blocks/index.js
@@ -1,7 +1,9 @@
 import CampaignEditor from './CampaignEditor';
 import CampaignProgressBar from './CampaignProgressBar';
+import CampaignFeaturedImage from './CampaignFeaturedImage';
 
 export {
 	CampaignEditor,
 	CampaignProgressBar,
+	CampaignFeaturedImage,
 };